[Chore] Extract pipeline manifest primitives
