+++ Cargo.toml
@@ -6,7 +6,7 @@ description = "Rust service for Haruki Sekai Asset Updater"
 license = "MIT"
 
 [workspace]
-members = ["."]
+members = [".", "crates/sekai-asset-pipeline"]
 resolver = "2"
 
 [dependencies]
@@ -33,6 +33,7 @@ reqwest = { version = "0.13", default-features = false, features = ["brotli", "d
 rmp-serde = "1.3"
 rsmpeg = { version = "0.18.0", default-features = false, features = ["ffmpeg7_1", "link_system_ffmpeg"], optional = true }
 serde = { version = "1.0", features = ["derive"] }
+sekai-asset-pipeline = { path = "crates/sekai-asset-pipeline" }
 sha2 = "0.11.0"
 sonic-rs = "0.5"
 tempfile = "3.27"
