+++ crates/sekai-asset-pipeline/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "sekai-asset-pipeline"
+version = "0.1.0"
+edition = "2021"
+description = "Reusable single-bundle primitives for Sekai asset pipelines"
+license = "MIT"
+publish = false
+
+[dependencies]
+aes = "0.9"
+cbc = { version = "0.2", features = ["alloc"] }
+hex = "0.4"
+rmp-serde = "1.3"
+serde = { version = "1.0", features = ["derive"] }
+thiserror = "2.0"
